refactor: replace standalone language switcher with modal-based selection in user preferences
- Create reusable LanguageSelectionModal component integrated with i18n system
- Replace hardcoded language list in UserPreference with actual i18n languages
- Remove standalone LanguageSwitcher component and its dedicated section
- Add missing translation keys for better i18n support
- Consolidate language selection into user preferences for better UX
