+++ app/(tabs)/profile/index.tsx
@@ -14,7 +14,6 @@ import { MaterialCommunityIcons } from "@expo/vector-icons";
 import { useRouter } from "expo-router";
 import PreferencesSection from "@/components/profile/UserPreference";
 import { useTranslation } from "react-i18next";
-import LanguageSwitcher from "@/components/shared/LanguageSwitcher";
 import { useAuthAndUser } from "@/hooks";
 import { Child, UserProfile } from "@/types";
 
@@ -193,10 +192,6 @@ export default function ProfileScreen() {
         {/* Children Information - Only show for signed in users */}
         {isAuthenticated && <ChildrenSection />}
 
-        {/* Language Switcher */}
-        <View style={styles.section}>
-          <LanguageSwitcher />
-        </View>
 
         {/* User Preferences */}
         <PreferencesSection />
