+++ components/profile/UserPreference.tsx
@@ -5,49 +5,27 @@ import {
   StyleSheet,
   Switch,
   TouchableOpacity,
-  FlatList,
-  Modal,
 } from "react-native";
 
 import { MaterialCommunityIcons } from "@expo/vector-icons";
+import { useTranslation } from "react-i18next";
+import { getCurrentLanguage, getAvailableLanguages } from "@/locales/i18n";
+import LanguageSelectionModal from "@/components/shared/LanguageSelectionModal";
 
 const primaryColor = "#00B493";
 
-const languages = [
-  "English",
-  "Spanish",
-  "French",
-  "German",
-  "Chinese",
-  "Japanese",
-  "Korean",
-];
 
-// Optimized render function
-const createLanguageRenderItem =
-  (
-    language: string,
-    handleLanguageSelect: (lang: string) => void,
-    primaryColor: string,
-  ) =>
-  ({ item }: { item: string }) => (
-    <TouchableOpacity
-      style={styles.languageOption}
-      onPress={() => handleLanguageSelect(item)}
-    >
-      <Text style={styles.languageOptionText}>{item}</Text>
-      {language === item && (
-        <MaterialCommunityIcons name="check" size={20} color={primaryColor} />
-      )}
-    </TouchableOpacity>
-  );
 
 export default function PreferencesSection() {
+  const { t } = useTranslation('profile');
   const [pushNotifications, setPushNotifications] = useState(true);
   const [historySave, setHistorySave] = useState(true);
   const [translationSave, setTranslationSave] = useState(true);
-  const [language, setLanguage] = useState("English");
   const [showLanguageModal, setShowLanguageModal] = useState(false);
+
+  const currentLanguage = getCurrentLanguage();
+  const availableLanguages = getAvailableLanguages();
+  const currentLanguageDisplay = availableLanguages.find(lang => lang.code === currentLanguage)?.nativeName || currentLanguage;
 
   const handleSetPushNotifications = (value: boolean) => {
     setPushNotifications(value);
@@ -59,17 +37,10 @@ export default function PreferencesSection() {
     setTranslationSave(value);
   };
 
-  const handleLanguageSelect = (selectedLanguage: string) => {
-    setLanguage(selectedLanguage);
-    setShowLanguageModal(false);
+  const handleLanguageChanged = () => {
+    // Language change is handled by the modal component
+    // This callback is just for any additional logic if needed
   };
-
-  // Create render function with current state
-  const renderLanguageItem = createLanguageRenderItem(
-    language,
-    handleLanguageSelect,
-    primaryColor,
-  );
 
   return (
     <View style={styles.section}>
@@ -136,14 +107,14 @@ export default function PreferencesSection() {
           <View style={styles.preferenceInfo}>
             <MaterialCommunityIcons name="earth" size={20} color="#666" />
             <View style={styles.preferenceText}>
-              <Text style={styles.preferenceTitle}>Preferred Language</Text>
+              <Text style={styles.preferenceTitle}>{t('preferences.language')}</Text>
               <Text style={styles.preferenceSubtitle}>
-                Language for app interface
+                {t('preferences.languageDescription')}
               </Text>
             </View>
           </View>
           <View style={styles.languageDisplay}>
-            <Text style={styles.languageText}>{language}</Text>
+            <Text style={styles.languageText}>{currentLanguageDisplay}</Text>
             <MaterialCommunityIcons
               name="chevron-right"
               size={20}
@@ -154,30 +125,11 @@ export default function PreferencesSection() {
       </View>
 
       {/* Language Selection Modal */}
-      <Modal
+      <LanguageSelectionModal
         visible={showLanguageModal}
-        transparent={true}
-        animationType="slide"
-        onRequestClose={() => setShowLanguageModal(false)}
-      >
-        <View style={styles.modalOverlay}>
-          <View style={styles.languageModal}>
-            <View style={styles.languageModalHeader}>
-              <Text style={styles.languageModalTitle}>Select Language</Text>
-              <TouchableOpacity onPress={() => setShowLanguageModal(false)}>
-                <MaterialCommunityIcons name="close" size={24} color="#333" />
-              </TouchableOpacity>
-            </View>
-            <FlatList
-              data={languages}
-              renderItem={renderLanguageItem}
-              keyExtractor={(item) => item}
-              style={styles.languageList}
-              showsVerticalScrollIndicator={false}
-            />
-          </View>
-        </View>
-      </Modal>
+        onClose={() => setShowLanguageModal(false)}
+        onLanguageChanged={handleLanguageChanged}
+      />
     </View>
   );
 }
@@ -237,49 +189,6 @@ const styles = StyleSheet.create({
     gap: 8,
   },
   languageText: {
-    fontSize: 16,
-    color: "#333",
-  },
-  modalOverlay: {
-    flex: 1,
-    backgroundColor: "rgba(0, 0, 0, 0.5)",
-    justifyContent: "center",
-    alignItems: "center",
-    padding: 20,
-  },
-  languageModal: {
-    backgroundColor: "#fff",
-    borderRadius: 16,
-    width: "100%",
-    maxHeight: "70%",
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 4 },
-    shadowOpacity: 0.25,
-    shadowRadius: 16,
-    elevation: 10,
-  },
-  languageModalHeader: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    padding: 20,
-    borderBottomWidth: 1,
-    borderBottomColor: "#e0e0e0",
-  },
-  languageModalTitle: {
-    fontSize: 20,
-    fontWeight: "600",
-    color: "#333",
-  },
-  languageOption: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    padding: 16,
-    borderBottomWidth: 1,
-    borderBottomColor: "#f0f0f0",
-  },
-  languageOptionText: {
     fontSize: 16,
     color: "#333",
   },
