+++ locales/ko/profile.json
@@ -35,6 +35,7 @@
   "preferences": {
     "title": "환경 설정",
     "language": "언어",
+    "languageDescription": "앱 인터페이스 언어",
     "selectLanguage": "언어 선택",
     "notifications": {
       "title": "알림",
